make tracker module Jira project setup helper a conftest fixture
